changes for graph.js debug
